client app

+++ backend/controllers/goalControllers.js
@@ -1,23 +1,25 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 // @desc        get goals
 // @route       GET /api/goals
 // @access      private
 const getGoals = asyncHandler(async (req, res) => {
     try {
-        const goals = await Goal.find();
-        res.status(200).json(goals);        
+        const goals = req.user.isAdmin
+            ? await Goal.find()
+            : await Goal.find({ user: req.user.id });
+        res.status(200).json(goals);
     } catch (error) {
         throw error;
     }
-})
+});
 
 // @desc        set goal
 // @route       POST /api/goals
 // @access      private
 const setGoal = asyncHandler(async (req, res) => {
-    console.log(req.body);
     if (!req.body.title) {
         // return res.status(400).json({ message: "Enter title of goal" });
         res.status(400);
@@ -31,13 +33,14 @@ const setGoal = asyncHandler(async (req, res) => {
     try {
         const goal = await Goal.create({
             title: req.body.title,
-            description: req.body.description
+            description: req.body.description,
+            user: req.user.id,
         });
         res.status(200).json(goal);
     } catch (error) {
         throw error;
     }
-})
+});
 
 // @desc        get goals
 // @route       PUT /api/goals/:id
@@ -49,12 +52,27 @@ const updateGoal = asyncHandler(async (req, res) => {
             res.status(404);
             throw new Error("Goal not found!");
         }
-        const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, { new: true });
-        res.status(200).json(updatedGoal);        
+        // check user
+        const user = await User.findById(req.user.id);
+        if (!user) {
+            res.status(401);
+            throw new Error("User not found!");
+        }
+        // make sure logged in user matches the goal user
+        if (goal.user.toString() !== user.id && !user.isAdmin) {
+            res.status(401);
+            throw new Error("User not authorized!");
+        }
+        const updatedGoal = await Goal.findByIdAndUpdate(
+            req.params.id,
+            req.body,
+            { new: true }
+        );
+        res.status(200).json(updatedGoal);
     } catch (error) {
         throw error;
     }
-})
+});
 
 // @desc        get goals
 // @route       DELETE /api/goals/:id
@@ -66,11 +84,22 @@ const deleteGoal = asyncHandler(async (req, res) => {
             res.status(404);
             throw new Error("Goal not found!");
         }
+        // check user
+        const user = await User.findById(req.user.id);
+        if (!user) {
+            res.status(401);
+            throw new Error("User not found!");
+        }
+        // make sure logged in user matches the goal user
+        if (goal.user.toString() !== user.id && !user.isAdmin) {
+            res.status(401);
+            throw new Error("User not authorized!");
+        }
         await goal.deleteOne();
-        res.status(200).json({ id: req.params.id });        
+        res.status(200).json({ id: req.params.id });
     } catch (error) {
         throw error;
     }
-})
+});
 
 module.exports = { getGoals, setGoal, updateGoal, deleteGoal };
